+++ src/Silex/RedirectableUrlMatcher.php
@@ -21,76 +21,41 @@ use Symfony\Component\Routing\Matcher\RedirectableUrlMatcher as BaseRedirectable
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-// symfony/routing: >=6.0.0
-if (!class_exists('\Symfony\Component\Routing\RouteCollectionBuilder')) {
-    class RedirectableUrlMatcher extends BaseRedirectableUrlMatcher
+class RedirectableUrlMatcher extends BaseRedirectableUrlMatcher
+{
+    use RedirectableUrlMatcherTrait;
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function executeRedirect(string $path, string $route, string $scheme = null): array
     {
-        /**
-         * {@inheritdoc}
-         */
-        public function redirect(string $path, string $route, string $scheme = null): array
-        {
-            $url = $this->context->getBaseUrl().$path;
-            $query = $this->context->getQueryString() ?: '';
+        $url = $this->context->getBaseUrl() . $path;
+        $query = $this->context->getQueryString() ?: '';
 
-            if ($query !== '') {
-                $url .= '?'.$query;
-            }
-
-            if ($this->context->getHost()) {
-                if ($scheme) {
-                    $port = '';
-                    if ('http' === $scheme && 80 != $this->context->getHttpPort()) {
-                        $port = ':'.$this->context->getHttpPort();
-                    } elseif ('https' === $scheme && 443 != $this->context->getHttpsPort()) {
-                        $port = ':'.$this->context->getHttpsPort();
-                    }
-
-                    $url = $scheme.'://'.$this->context->getHost().$port.$url;
-                }
-            }
-
-            return array(
-                '_controller' => function ($url) { return new RedirectResponse($url, 301); },
-                '_route' => null,
-                'url' => $url,
-            );
+        if ($query !== '') {
+            $url .= '?' . $query;
         }
-    }
-} else {
-    class RedirectableUrlMatcher extends BaseRedirectableUrlMatcher
-    {
-        /**
-         * {@inheritdoc}
-         */
-        public function redirect($path, $route, $scheme = null)
-        {
-            $url = $this->context->getBaseUrl().$path;
-            $query = $this->context->getQueryString() ?: '';
 
-            if ($query !== '') {
-                $url .= '?'.$query;
-            }
-
-            if ($this->context->getHost()) {
-                if ($scheme) {
-                    $port = '';
-                    if ('http' === $scheme && 80 != $this->context->getHttpPort()) {
-                        $port = ':'.$this->context->getHttpPort();
-                    } elseif ('https' === $scheme && 443 != $this->context->getHttpsPort()) {
-                        $port = ':'.$this->context->getHttpsPort();
-                    }
-
-                    $url = $scheme.'://'.$this->context->getHost().$port.$url;
+        if ($this->context->getHost()) {
+            if ($scheme) {
+                $port = '';
+                if ('http' === $scheme && 80 != $this->context->getHttpPort()) {
+                    $port = ':' . $this->context->getHttpPort();
+                } elseif ('https' === $scheme && 443 != $this->context->getHttpsPort()) {
+                    $port = ':' . $this->context->getHttpsPort();
                 }
-            }
 
-            return array(
-                '_controller' => function ($url) { return new RedirectResponse($url, 301); },
-                '_route' => null,
-                'url' => $url,
-            );
+                $url = $scheme . '://' . $this->context->getHost() . $port . $url;
+            }
         }
+
+        return array(
+            '_controller' => function ($url) {
+                return new RedirectResponse($url, 301);
+            },
+            '_route' => null,
+            'url' => $url,
+        );
     }
 }
-
